+++ src/App.js
@@ -1,10 +1,10 @@
 import './App.css';
-import MainVisual from './components/MainVisual';
+import AppContainer from './components/AppContainer';
 
 function App() {
   return (
     <div className="App">
-      <MainVisual />
+      <AppContainer />
     </div>
   );
 }
